# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Gofin API
  description: >
    Personal finance management API.
    All protected endpoints require a JWT bearer token in the Authorization header.
    All ID fields use UUID v7 format.
  version: "1.0.0"
  contact:
    name: Gofin
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OAuthProvider:
      name: provider
      in: path
      required: true
      schema:
        type: string
      description: OAuth provider name
  schemas:
    Error:
      type: object
      properties:
        status:
          type: integer
          example: 400
        title:
          type: string
          example: Bad Request
        detail:
          type: string
          example: Invalid request body
...
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          example: 3600
        token_type:
          type: string
          example: Bearer
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
security:
  - BearerAuth: []
paths:
...
  /api/v1/auth/refresh:
    post:
      summary: Refresh token
      description: Exchange a refresh token for a new access token.
      operationId: refreshToken
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/auth/{provider}/url:
    get:
      summary: OAuth authorization URL
      description: Get the OAuth authorization URL for the specified provider.
      operationId: oauthURL
      tags: [Auth]
      security: []
      parameters:
        - $ref: "#/components/parameters/OAuthProvider"
      responses:
        "200":
          description: OAuth authorization URL
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/auth/{provider}/callback:
    get:
      summary: OAuth callback
      description: Handle OAuth callback from the specified provider.
      operationId: oauthCallback
      tags: [Auth]
      security: []
      parameters:
        - $ref: "#/components/parameters/OAuthProvider"
      responses:
        "200":
          description: OAuth authentication result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # -- Users ----------------------------------------------------------------
...